+++ src/index.js
@@ -4,8 +4,7 @@ import "./css/icons.scss"
 import demo from "./demo";
 import getRankChange from "./rank-change";
 import getTeamIcons from "./team-icons";
-import shishenTable from "../static/json/shishen.json"
-import rankTable from "../static/json/shishen_rank.json"
+import initShishenRank from "./init/shishen-rank";
 
 $(window).on('load', function () {
     console.log("window done");
@@ -15,28 +14,6 @@ $(window).on('load', function () {
     showUpdates();
     demo();
 });
-
-
-function initShishenRank() {
-    var shishen_rank = [];
-    for (let key in rankTable) {
-        let shishen_html = getTeamIcons([rankTable[key][2]]);
-        shishen_html = shishen_html + shishenTable[rankTable[key][2]]['name']
-        let win_rate_html = createProgBar(rankTable[key][3] / 100, 1);
-        let pick_rate_html = createProgBar(rankTable[key][4], 1);
-        let rank_change = getRankChange(rankTable[key][1]);
-        let row = {
-            rank: rankTable[key][0],
-            rank_change: rank_change,
-            shishen: shishen_html,
-            win_rate: win_rate_html,
-            pick_rate: pick_rate_html,
-            total_matches: rankTable[key][5]
-        };
-        shishen_rank.push(row)
-    }
-    $('#shishen-rank-table').bootstrapTable('load', shishen_rank);
-}
 
 
 // 注册给遗留系统
